move loop to metainfo package
